refactor(runner): split start sandbox finalization
